requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
pyarrow
requests
httpx
sqlalchemy
psycopg2-binary
dagster
dagster-webserver
